+++ src/tags/interfaces/tags.interface.ts
@@ -1,0 +1,3 @@
+export interface TagsResponse {
+  tags: string[];
+}
